requirements.txt
# Apache Airflow và plugins
apache-airflow==2.9.0
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-snowflake==5.1.1
apache-airflow-providers-apache-spark==4.3.0
apache-airflow-providers-slack==8.1.0  # cho alerting

# Database connections
psycopg2-binary==2.9.9  # PostgreSQL connector
snowflake-connector-python==3.5.0
sqlalchemy==1.4.50

# Data processing
pyspark==3.5.0
delta-spark==3.0.0  # nếu muốn dùng Delta Lake format
pandas==2.1.4
numpy==1.24.3

# MinIO
minio==7.2.0

# dbt
dbt-core==1.7.3
dbt-postgres==1.7.3  # để test locally
dbt-snowflake==1.7.3
